# docs/openapi/auto-edit.swagger.yaml
openapi: 3.0.3
info:
  title: FireRed OpenStoryline Auto Edit API
  version: 1.0.0
  description: |
    自动化视频剪辑接口文档。

    该 Swagger 文档基于 FireRed-OpenStoryline 当前会话、素材上传、Agent 与 MCP/Node 剪辑能力设计，
    用于描述“创建剪辑会话、上传素材、提交自动剪辑任务、获取剪辑结果”四个接口。

    注意：本文档用于接口交互设计与调试说明，不修改项目现有服务逻辑。

servers:
  - url: http://localhost:7860
    description: Local development server
  - url: http://127.0.0.1:7860
    description: Local development server by IP

tags:
  - name: Sessions
    description: 剪辑会话
  - name: Media
    description: 素材上传
  - name: Auto Edit
    description: 自动剪辑任务

paths:
  /api/sessions:
    post:
      tags:
        - Sessions
      summary: 创建剪辑会话
      description: 创建一个新的自动剪辑会话，后续素材上传和剪辑任务都基于该会话执行。
      operationId: createEditSession
      responses:
        "200":
          description: 会话创建成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateSessionResponse"
              example:
                session_id: abc123
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /api/sessions/{session_id}/media:
    post:
      tags:
        - Media
      summary: 上传素材
      description: |
        上传视频、音频或图片素材。上传完成后返回 `media_id`，自动剪辑任务通过 `media_id` 引用素材。

        支持的素材类型：
        - 视频：mp4、mov、mkv、avi 等
        - 音频：mp3、wav、m4a、aac 等
        - 图片：jpg、jpeg、png、webp、bmp 等
      operationId: uploadEditMedia
      parameters:
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: 上传的视频、音频或图片文件。
      responses:
        "200":
          description: 素材上传成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadMediaResponse"
              example:
                media_id: media_001
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/SessionNotFound"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /api/sessions/{session_id}/edit:
    post:
      tags:
        - Auto Edit
      summary: 提交自动剪辑任务
      description: |
        提交自动剪辑任务。系统会根据用户文案、剪辑要求、素材 ID 和必要上下文，由 Agent 自动分析剪辑流程，
        并调用 MCP / Node 执行素材分析、裁剪、混剪、字幕、BGM、图片叠加、音频处理、编码等操作。

        该接口返回 `processing` 表示任务已进入后台处理流程，客户端应通过
        `GET /api/sessions/{session_id}/result` 轮询任务状态。
      operationId: submitAutoEditTask
      parameters:
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AutoEditRequest"
            example:
              text: 制作一个产品宣传视频
              requirement: 将两个视频混剪，添加中文字幕和背景音乐，控制在30秒以内
              media_ids:
                - media_001
                - media_002
      responses:
        "200":
          description: 任务已提交
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmitEditResponse"
              example:
                status: processing
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/SessionNotFound"
        "409":
          description: 当前会话已有剪辑任务正在处理
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                detail: current edit task is still processing
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /api/sessions/{session_id}/result:
    get:
      tags:
        - Auto Edit
      summary: 获取剪辑结果
      description: |
        查询当前会话的自动剪辑任务状态。

        当状态为 `completed` 时，返回最终视频访问地址 `video_url`。
        客户端可继续请求该地址下载或播放剪辑完成的视频。
      operationId: getAutoEditResult
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EditResultResponse"
              examples:
                processing:
                  summary: 任务处理中
                  value:
                    status: processing
                completed:
                  summary: 任务完成
                  value:
                    status: completed
                    video_url: /api/sessions/abc123/result.mp4
                failed:
                  summary: 任务失败
                  value:
                    status: failed
                    error: render_video failed
        "404":
          $ref: "#/components/responses/SessionNotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"

components:
  parameters:
    SessionId:
      name: session_id
      in: path
      required: true
      description: 当前剪辑会话唯一标识。
      schema:
        type: string
      example: abc123

  schemas:
    CreateSessionResponse:
      type: object
      required:
        - session_id
      properties:
        session_id:
          type: string
          description: 当前剪辑会话唯一标识。
          example: abc123

    UploadMediaResponse:
      type: object
      required:
        - media_id
      properties:
        media_id:
          type: string
          description: 上传素材的唯一标识，后续剪辑任务通过该 ID 引用素材。
          example: media_001

    AutoEditRequest:
      type: object
      required:
        - text
        - requirement
        - media_ids
      properties:
        text:
          type: string
          minLength: 1
          description: 用户想要实现的视频内容或文案。
          example: 制作一个介绍智能营销终端的30秒产品宣传视频。
        requirement:
          type: string
          minLength: 1
          description: 具体剪辑要求，例如字幕、BGM、混剪、镜头排序、时长控制等。
          example: 将两个视频进行混剪，控制在30秒以内，添加中文字幕，并添加背景音乐。
        media_ids:
          type: array
          minItems: 1
          description: 当前剪辑任务需要使用的素材 ID 列表。
          items:
            type: string
          example:
            - media_001
            - media_002

    SubmitEditResponse:
      type: object
      required:
        - status
      properties:
        status:
          $ref: "#/components/schemas/EditStatus"

    EditResultResponse:
      type: object
      required:
        - status
      properties:
        status:
          $ref: "#/components/schemas/EditStatus"
        video_url:
          type: string
          nullable: true
          description: 最终视频访问地址，仅当 `status=completed` 时返回。
          example: /api/sessions/abc123/result.mp4
        error:
          type: string
          nullable: true
          description: 任务失败原因，仅当 `status=failed` 时返回。
          example: render_video failed

    EditStatus:
      type: string
      description: 当前剪辑任务状态。
      enum:
        - idle
        - processing
        - completed
        - failed
      example: processing

    ErrorResponse:
      type: object
      properties:
        detail:
          oneOf:
            - type: string
            - type: object
            - type: array
              items:
                type: object
          description: 错误详情。
          example: session not found

  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            detail: invalid request

    SessionNotFound:
      description: 会话不存在
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            detail: session not found

    PayloadTooLarge:
      description: 上传文件过大
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            detail: file too large

    UnsupportedMediaType:
      description: 不支持的素材类型
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            detail: unsupported media type

    ValidationError:
      description: 请求体校验失败
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

    TooManyRequests:
      description: 请求过于频繁
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            detail: rate limited

    InternalServerError:
      description: 服务内部错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            detail: internal server error